+++ docs/changelog.rst
@@ -4,5 +4,6 @@ Change Log
 v0.1 (Not yet released)
 -----------------------
 
+* ``easy_shell``: ``exorlink`` (distance 2 to 6).
 * Initial release